refactor: extract tabsheet-mixin
