+++ packages/tabsheet/package.json
@@ -35,6 +35,7 @@
     "polymer"
   ],
   "dependencies": {
+    "@open-wc/dedupe-mixin": "^1.3.0",
     "@polymer/polymer": "^3.0.0",
     "@vaadin/component-base": "24.4.0-alpha0",
     "@vaadin/scroller": "24.4.0-alpha0",
