+++ packages/tabsheet/src/vaadin-tabsheet.d.ts
@@ -8,6 +8,7 @@ import { DelegateStateMixin } from '@vaadin/component-base/src/delegate-state-mi
 import { ElementMixin } from '@vaadin/component-base/src/element-mixin.js';
 import type { Tab } from '@vaadin/tabs/src/vaadin-tab.js';
 import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
+import { TabSheetMixin } from './vaadin-tabsheet-mixin.js';
 
 /**
  * Fired when the `items` property changes.
@@ -69,21 +70,9 @@ export interface TabSheetEventMap extends HTMLElementEventMap, TabSheetCustomEve
  * @fires {CustomEvent} items-changed - Fired when the `items` property changes.
  * @fires {CustomEvent} selected-changed - Fired when the `selected` property changes.
  */
-declare class TabSheet extends ControllerMixin(DelegateStateMixin(ElementMixin(ThemableMixin(HTMLElement)))) {
-  /**
-   * The index of the selected tab.
-   */
-  selected: number | null | undefined;
-
-  /**
-   * The list of `<vaadin-tab>`s from which a selection can be made.
-   * It is populated from the elements passed inside the slotted
-   * `<vaadin-tabs>`, and updated dynamically when adding or removing items.
-   *
-   * Note: unlike `<vaadin-combo-box>`, this property is read-only.
-   */
-  readonly items: Tab[] | undefined;
-
+declare class TabSheet extends TabSheetMixin(
+  ControllerMixin(DelegateStateMixin(ElementMixin(ThemableMixin(HTMLElement)))),
+) {
   addEventListener<K extends keyof TabSheetEventMap>(
     type: K,
     listener: (this: TabSheet, ev: TabSheetEventMap[K]) => void,
